separated movement logic into specialized manager
