--- Assets/Engine/Board/Scripts/Editor.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: c8dfd956672b99c49b89e2ebbe67925a
-folderAsset: yes
-timeCreated: 1500016207
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
