+++ Assets/Engine/Game/Scripts/Rules/WARControlMovement.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 50e22a94d90ba4737bb62fa86a5cad2d
+timeCreated: 1501397766
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
