--- Assets/Resources/Armies/Shmoogaloo/HQ.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: e7d4bb26f2bb409419e34858002fa3c4
-folderAsset: yes
-timeCreated: 1499583520
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
