+++ Assets/Resources/Managers/UIInput.prefab
@@ -9,17 +9,17 @@ Prefab:
     m_Modifications: []
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 0}
-  m_RootGameObject: {fileID: 1613480557772346}
+  m_RootGameObject: {fileID: 1764169965864502}
   m_IsPrefabParent: 1
---- !u!1 &1613480557772346
+--- !u!1 &1764169965864502
 GameObject:
   m_ObjectHideFlags: 0
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   serializedVersion: 5
   m_Component:
-  - component: {fileID: 4138575522782364}
-  - component: {fileID: 114710390323066008}
+  - component: {fileID: 4889065710913040}
+  - component: {fileID: 114295117866412876}
   m_Layer: 0
   m_Name: UIInput
   m_TagString: Untagged
@@ -27,12 +27,12 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &4138575522782364
+--- !u!4 &4889065710913040
 Transform:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1613480557772346}
+  m_GameObject: {fileID: 1764169965864502}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
@@ -40,12 +40,12 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &114710390323066008
+--- !u!114 &114295117866412876
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1613480557772346}
+  m_GameObject: {fileID: 1764169965864502}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: a589417b7089ede4ba442907a577f571, type: 3}
