+++ Assets/Resources/Managers/UIInput.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
 guid: 6205fecfbf6cad548a667495cce6eacf
-timeCreated: 1501132168
+timeCreated: 1501398036
 licenseType: Pro
 NativeFormatImporter:
   mainObjectFileID: 100100000
