+++ Assets/Resources/Managers/WARControlBoard.prefab
@@ -9,17 +9,17 @@ Prefab:
     m_Modifications: []
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 0}
-  m_RootGameObject: {fileID: 1425572701221682}
+  m_RootGameObject: {fileID: 1197937367482134}
   m_IsPrefabParent: 1
---- !u!1 &1425572701221682
+--- !u!1 &1197937367482134
 GameObject:
   m_ObjectHideFlags: 0
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   serializedVersion: 5
   m_Component:
-  - component: {fileID: 4010607718803336}
-  - component: {fileID: 114131144603952294}
+  - component: {fileID: 4437924090333468}
+  - component: {fileID: 114701978266092338}
   m_Layer: 0
   m_Name: WARControlBoard
   m_TagString: Untagged
@@ -27,12 +27,12 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &4010607718803336
+--- !u!4 &4437924090333468
 Transform:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1425572701221682}
+  m_GameObject: {fileID: 1197937367482134}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
@@ -40,12 +40,12 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &114131144603952294
+--- !u!114 &114701978266092338
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1425572701221682}
+  m_GameObject: {fileID: 1197937367482134}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: b961bc3d577e0fd488b3cfb9c5b181eb, type: 3}
