+++ Assets/Resources/Managers/WARControlBoard.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
 guid: fffb61bfa1919ff4fb7d04aa28c5fcb1
-timeCreated: 1501132166
+timeCreated: 1501398035
 licenseType: Pro
 NativeFormatImporter:
   mainObjectFileID: 100100000
