+++ Assets/Resources/Managers/WARControlDeployment.prefab
@@ -9,17 +9,17 @@ Prefab:
     m_Modifications: []
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 0}
-  m_RootGameObject: {fileID: 1889673606027330}
+  m_RootGameObject: {fileID: 1206294278434280}
   m_IsPrefabParent: 1
---- !u!1 &1889673606027330
+--- !u!1 &1206294278434280
 GameObject:
   m_ObjectHideFlags: 0
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   serializedVersion: 5
   m_Component:
-  - component: {fileID: 4310844148114054}
-  - component: {fileID: 114817986629286568}
+  - component: {fileID: 4069128959512342}
+  - component: {fileID: 114281166471168776}
   m_Layer: 0
   m_Name: WARControlDeployment
   m_TagString: Untagged
@@ -27,12 +27,12 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &4310844148114054
+--- !u!4 &4069128959512342
 Transform:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1889673606027330}
+  m_GameObject: {fileID: 1206294278434280}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
@@ -40,12 +40,12 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &114817986629286568
+--- !u!114 &114281166471168776
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1889673606027330}
+  m_GameObject: {fileID: 1206294278434280}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: aca005fc522d6b1448b3cc23b62f4527, type: 3}
