+++ Assets/Resources/Managers/WARControlMovement.prefab
@@ -9,30 +9,30 @@ Prefab:
     m_Modifications: []
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 0}
-  m_RootGameObject: {fileID: 1262706514485704}
+  m_RootGameObject: {fileID: 1073361837751982}
   m_IsPrefabParent: 1
---- !u!1 &1262706514485704
+--- !u!1 &1073361837751982
 GameObject:
   m_ObjectHideFlags: 0
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   serializedVersion: 5
   m_Component:
-  - component: {fileID: 4133763023875728}
-  - component: {fileID: 114490995439712546}
+  - component: {fileID: 4149902734046288}
+  - component: {fileID: 114986468138070662}
   m_Layer: 0
-  m_Name: WARControlGameplay
+  m_Name: WARControlMovement
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &4133763023875728
+--- !u!4 &4149902734046288
 Transform:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1262706514485704}
+  m_GameObject: {fileID: 1073361837751982}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
@@ -40,15 +40,14 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &114490995439712546
+--- !u!114 &114986468138070662
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1262706514485704}
+  m_GameObject: {fileID: 1073361837751982}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5fb08f7f37eba914ab74b5235f329507, type: 3}
+  m_Script: {fileID: 11500000, guid: 50e22a94d90ba4737bb62fa86a5cad2d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  currentPlayer: 0
