+++ Assets/Resources/Managers/WARControlMovement.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: ae35babb5601aa1428c81fa71038aa55
-timeCreated: 1501132167
+guid: 37582d187a68943f88d6771cb4aa2747
+timeCreated: 1501398035
 licenseType: Pro
 NativeFormatImporter:
   mainObjectFileID: 100100000
