+++ Assets/Resources/Managers/WARControlSelection.prefab
@@ -9,17 +9,17 @@ Prefab:
     m_Modifications: []
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 0}
-  m_RootGameObject: {fileID: 1450474512700642}
+  m_RootGameObject: {fileID: 1353049224607362}
   m_IsPrefabParent: 1
---- !u!1 &1450474512700642
+--- !u!1 &1353049224607362
 GameObject:
   m_ObjectHideFlags: 0
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   serializedVersion: 5
   m_Component:
-  - component: {fileID: 4081289270053496}
-  - component: {fileID: 114018634705929394}
+  - component: {fileID: 4261004493105946}
+  - component: {fileID: 114252680128784556}
   m_Layer: 0
   m_Name: WARControlSelection
   m_TagString: Untagged
@@ -27,12 +27,12 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &4081289270053496
+--- !u!4 &4261004493105946
 Transform:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1450474512700642}
+  m_GameObject: {fileID: 1353049224607362}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
@@ -40,12 +40,12 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &114018634705929394
+--- !u!114 &114252680128784556
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1450474512700642}
+  m_GameObject: {fileID: 1353049224607362}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 0ed65d711723b90459cacb8d83e78984, type: 3}
