+++ Assets/Resources/Managers/WARControlSetup.prefab
@@ -9,17 +9,17 @@ Prefab:
     m_Modifications: []
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 0}
-  m_RootGameObject: {fileID: 1549856242424856}
+  m_RootGameObject: {fileID: 1290659826223354}
   m_IsPrefabParent: 1
---- !u!1 &1549856242424856
+--- !u!1 &1290659826223354
 GameObject:
   m_ObjectHideFlags: 0
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   serializedVersion: 5
   m_Component:
-  - component: {fileID: 4945254755514608}
-  - component: {fileID: 114583610412406944}
+  - component: {fileID: 4370476407983054}
+  - component: {fileID: 114577586619178810}
   m_Layer: 0
   m_Name: WARControlSetup
   m_TagString: Untagged
@@ -27,12 +27,12 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &4945254755514608
+--- !u!4 &4370476407983054
 Transform:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1549856242424856}
+  m_GameObject: {fileID: 1290659826223354}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
@@ -40,12 +40,12 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &114583610412406944
+--- !u!114 &114577586619178810
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1549856242424856}
+  m_GameObject: {fileID: 1290659826223354}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 3413b2f65d1163549b0b62d2944df09f, type: 3}
