+++ Assets/Resources/Managers/WARControlSetup.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
 guid: d71d345568fb6464ca60c5357e0449f3
-timeCreated: 1501132167
+timeCreated: 1501398036
 licenseType: Pro
 NativeFormatImporter:
   mainObjectFileID: 100100000
